requirements.txt
fastapi[standard]>=0.112.0,<0.113.0
gitpython~=3.1.43
python-decouple~=3.8
requests-oauthlib~=2.0.0
loguru==0.7.2
motor~=3.5.1
PyJWT~=2.9.0
